requirements.txt
Django>=4.2.0,<5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-extensions>=3.2.0
Pillow>=9.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
django-filter>=23.0
django-tables2>=2.6.0
python-dateutil>=2.8.0
